removed useless casts to long

+++ gwt-nio/src/main/java/java/nio/ByteBuffer.java
@@ -111,7 +111,7 @@ public abstract class ByteBuffer extends Buffer implements Comparable<ByteBuffer
      */
     public static ByteBuffer wrap(byte[] array, int start, int len) {
         int length = array.length;
-        if ((start < 0) || (len < 0) || ((long) start + (long) len > length)) {
+        if ((start < 0) || (len < 0) || (start + len > length)) {
             throw new IndexOutOfBoundsException();
         }
 
@@ -461,7 +461,7 @@ public abstract class ByteBuffer extends Buffer implements Comparable<ByteBuffer
      */
     public ByteBuffer get(byte[] dest, int off, int len) {
         int length = dest.length;
-        if ((off < 0 ) || (len < 0) || ((long)off + (long)len > length)) {
+        if ((off < 0 ) || (len < 0) || (off + len > length)) {
             throw new IndexOutOfBoundsException();
         }
 
@@ -837,7 +837,7 @@ public abstract class ByteBuffer extends Buffer implements Comparable<ByteBuffer
      */
     public ByteBuffer put(byte[] src, int off, int len) {
         int length = src.length;
-        if ((off < 0 ) || (len < 0) || ((long)off + (long)len > length)) {
+        if ((off < 0 ) || (len < 0) || (off + len > length)) {
             throw new IndexOutOfBoundsException();
         }
