+++ gwt-nio/src/main/java/java/nio/CharSequenceAdapter.java
@@ -72,7 +72,7 @@ final class CharSequenceAdapter extends CharBuffer {
 
     public final CharBuffer get(char[] dest, int off, int len) {
         int length = dest.length;
-        if ((off < 0 ) || (len < 0) || (long)off + (long)len > length) {
+        if ((off < 0 ) || (len < 0) || off + len > length) {
             throw new IndexOutOfBoundsException();
         }
         if (len > remaining()) {
@@ -117,7 +117,7 @@ final class CharSequenceAdapter extends CharBuffer {
     }
 
     public final CharBuffer put(char[] src, int off, int len) {
-        if ((off < 0 ) || (len < 0) || (long)off + (long)len > src.length) {
+        if ((off < 0 ) || (len < 0) || off + len > src.length) {
             throw new IndexOutOfBoundsException();
         }
         
@@ -129,7 +129,7 @@ final class CharSequenceAdapter extends CharBuffer {
     }
 
     public CharBuffer put(String src, int start, int end) {
-        if ((start < 0 ) || (end < 0) || (long)start + (long)end > src.length()) {
+        if ((start < 0 ) || (end < 0) || start + end > src.length()) {
             throw new IndexOutOfBoundsException();
         }
         throw new ReadOnlyBufferException();
