+++ gwt-nio/src/main/java/java/nio/ReadOnlyCharArrayBuffer.java
@@ -89,7 +89,7 @@ final class ReadOnlyCharArrayBuffer extends CharArrayBuffer {
     }
 
     public CharBuffer put(String src, int start, int end) {
-        if ((start < 0 ) || (end < 0) || (long)start + (long)end > src.length()) {
+        if ((start < 0 ) || (end < 0) || start + end > src.length()) {
             throw new IndexOutOfBoundsException();
         }
         throw new ReadOnlyBufferException();
