+++ gwt-nio/src/main/java/java/nio/ReadWriteHeapByteBuffer.java
@@ -110,7 +110,7 @@ final class ReadWriteHeapByteBuffer extends HeapByteBuffer {
      * @see java.nio.ByteBuffer#put(byte[], int, int)
      */
     public ByteBuffer put(byte[] src, int off, int len) {
-        if (off < 0 || len < 0 || (long)off + (long)len > src.length) {
+        if (off < 0 || len < 0 || off + len > src.length) {
             throw new IndexOutOfBoundsException();
         }
         if (len > remaining()) {
@@ -152,7 +152,7 @@ final class ReadWriteHeapByteBuffer extends HeapByteBuffer {
     }
 
     public ByteBuffer putInt(int index, int value) {
-        if (index < 0 || (long)index + 4 > limit) {
+        if (index < 0 || index + 4 > limit) {
             throw new IndexOutOfBoundsException();
         }
         store(index, value);
@@ -160,7 +160,7 @@ final class ReadWriteHeapByteBuffer extends HeapByteBuffer {
     }
 
     public ByteBuffer putLong(int index, long value) {
-        if (index < 0 || (long)index + 8 > limit) {
+        if (index < 0 || index + 8 > limit) {
             throw new IndexOutOfBoundsException();
         }
         store(index, value);
@@ -178,7 +178,7 @@ final class ReadWriteHeapByteBuffer extends HeapByteBuffer {
     }
 
     public ByteBuffer putShort(int index, short value) {
-        if (index < 0 || (long)index + 2 > limit) {
+        if (index < 0 || index + 2 > limit) {
             throw new IndexOutOfBoundsException();
         }
         store(index, value);
